Replace SetMixin by MutableSet
* ValueError and KeyError apparently quoted differently
* MutableSet requires __len__ and __contains__
* MutableSet wants discard (no throw) instead of remove (throws if missing)
* MutableSet does not provide update, even though it provides __ior__...

Asides

* removed not-really-necessary list creation from CheckboxValue.__iter__
* replaced KeyError by ValueError in CheckboxValue.add to match
  MultipleSelectOptions
