+++ src/lxml/html/__init__.py
@@ -35,11 +35,10 @@ import copy
 import sys
 import re
 
-from collections import MutableMapping
+from collections import MutableMapping, MutableSet
 
 from lxml import etree
 from lxml.html import defs
-from lxml.html._setmixin import SetMixin
 
 try:
     from urlparse import urljoin
@@ -1209,7 +1208,7 @@ class SelectElement(InputMixin, HtmlElement):
 
 HtmlElementClassLookup._default_element_classes['select'] = SelectElement
 
-class MultipleSelectOptions(SetMixin):
+class MultipleSelectOptions(MutableSet):
     """
     Represents all the selected options in a ``<select multiple>`` element.
 
@@ -1227,6 +1226,9 @@ class MultipleSelectOptions(SetMixin):
         return iter(_options_xpath(self.select))
     options = property(options)
 
+    def __contains__(self, item):
+        return any(has_item == item for has_item in self)
+
     def __iter__(self):
         for option in self.options:
             if 'selected' in option.attrib:
@@ -1236,6 +1238,9 @@ class MultipleSelectOptions(SetMixin):
                 if opt_value:
                     opt_value = opt_value.strip()
                 yield opt_value
+
+    def __len__(self):
+        return sum(1 for option in self.options if 'selected' in option.attrib)
 
     def add(self, item):
         for option in self.options:
@@ -1251,7 +1256,7 @@ class MultipleSelectOptions(SetMixin):
             raise ValueError(
                 "There is no option with the value %r" % item)
 
-    def remove(self, item):
+    def discard(self, item):
         for option in self.options:
             opt_value = option.get('value')
             if opt_value is None:
@@ -1261,13 +1266,13 @@ class MultipleSelectOptions(SetMixin):
             if opt_value == item:
                 if 'selected' in option.attrib:
                     del option.attrib['selected']
-                else:
-                    raise ValueError(
-                        "The option %r is not currently selected" % item)
                 break
-        else:
-            raise ValueError(
-                "There is not option with the value %r" % item)
+
+    # missing from MutableSet?
+    def update(self, *others):
+        for other in others:
+            for item in other:
+                self.add(item)
 
     def __repr__(self):
         return '<%s {%s} for select name=%r>' % (
@@ -1366,7 +1371,7 @@ class CheckboxGroup(list):
         return '%s(%s)' % (
             self.__class__.__name__, list.__repr__(self))
 
-class CheckboxValues(SetMixin):
+class CheckboxValues(MutableSet):
 
     """
     Represents the values of the checked checkboxes in a group of
@@ -1377,10 +1382,17 @@ class CheckboxValues(SetMixin):
         self.group = group
 
     def __iter__(self):
-        return iter([
+        return (
             el.get('value')
             for el in self.group
-            if 'checked' in el.attrib])
+            if 'checked' in el.attrib
+        )
+
+    def __contains__(self, item):
+        return any(has_item == item for item in self)
+
+    def __len__(self):
+        return sum(1 for el in self.group if 'checked' in el.attrib)
 
     def add(self, value):
         for el in self.group:
@@ -1388,20 +1400,20 @@ class CheckboxValues(SetMixin):
                 el.set('checked', '')
                 break
         else:
-            raise KeyError("No checkbox with value %r" % value)
+            raise ValueError("No checkbox with value %r" % value)
 
-    def remove(self, value):
+    def discard(self, value):
         for el in self.group:
             if el.get('value') == value:
                 if 'checked' in el.attrib:
                     del el.attrib['checked']
-                else:
-                    raise KeyError(
-                        "The checkbox with value %r was already unchecked" % value)
                 break
-        else:
-            raise KeyError(
-                "No checkbox with value %r" % value)
+
+    # missing from MutableSet?
+    def update(self, *others):
+        for other in others:
+            for item in other:
+                self.add(item)
 
     def __repr__(self):
         return '<%s {%s} for checkboxes name=%r>' % (
